refactoring, move components to separate files
